+++ src/main/java/UMC/news/newsIntelligent/domain/member/dto/MemberResponseDto.java
@@ -1,6 +1,6 @@
 package UMC.news.newsIntelligent.domain.member.dto;
 
-import UMC.news.newsIntelligent.domain.member.Member;
+import UMC.news.newsIntelligent.domain.member.entity.Member;
 
 public record MemberResponseDto(Long id, String email) {
     public static MemberResponseDto from(Member member) {
